requirements.txt
fastapi>=0.110.0
pydantic>=2.6.3
uvicorn>=0.28.0
requests>=2.31.0
responses>=0.25.0 # unit testing requests
pytest>=8.1.1
httpx>=0.27.0
python-dotenv>=1.0.1